accelerating swift compilation

+++ Calculator/CalculatorApp.swift
@@ -37,11 +37,13 @@ struct CalculatorApp: App {
     var body: some Scene {
         WindowGroup {
             // a little hack to prevent that which background creeps up during device orientation chang rotation
-            let expandedDeviceSize: CGFloat = CGFloat(1.5) * max(UIScreen.main.bounds.size.width, UIScreen.main.bounds.size.height)
+            let expandedDeviceSize: CGFloat = 1.5 * max(UIScreen.main.bounds.size.width, UIScreen.main.bounds.size.height)
             iOSSize(brain: brain)
                 .statusBar(hidden: true)
                 .background(Rectangle()
-                                .frame(width: expandedDeviceSize, height: expandedDeviceSize, alignment: .center)
+                                .frame(width: expandedDeviceSize,
+                                       height: expandedDeviceSize,
+                                       alignment: .center)
                                 .foregroundColor(TE.appBackgroundColor)
                                 .ignoresSafeArea())
         }
